- New Card Design - Font loading
