+++ app/src/main/java/tcp_uno/view/MenuView.java
@@ -1,6 +1,8 @@
 package tcp_uno.view;
 
 import tcp_uno.AppState;
+import tcp_uno.components.Background;
+import tcp_uno.components.ImageButton;
 
 import static com.raylib.Jaylib.RAYWHITE;
 import static com.raylib.Raylib.*;
